RSWW-119 reservation-service: add logic to send webhook message when someone purchases an offer
