+++ offer-provider-service/src/main/java/org/microarchitecturovisco/offerprovider/controllers/OffersController.java
@@ -2,15 +2,11 @@ package org.microarchitecturovisco.offerprovider.controllers;
 
 import lombok.RequiredArgsConstructor;
 import org.microarchitecturovisco.offerprovider.domain.dto.OfferDto;
-import org.microarchitecturovisco.offerprovider.domain.dto.responses.TransportDto;
 import org.microarchitecturovisco.offerprovider.domain.responses.GetOfferDetailsResponseDto;
 import org.microarchitecturovisco.offerprovider.services.OffersService;
-import org.microarchitecturovisco.offerprovider.websockets.OfferProviderWebSocketHandler;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.web.bind.annotation.*;
-import reactor.core.publisher.Mono;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -86,17 +82,5 @@ public class OffersController {
         logger.info("Response: " + responseDto);
 
         return responseDto;
-    }
-
-    private final OfferProviderWebSocketHandler webSocketHandler;
-
-    @Scheduled(initialDelay = 5000, fixedDelay = 5000)
-    public void testBuyOffer() {
-        Logger logger = Logger.getLogger("OP WebSocket Producer");
-
-        UUID offerId = UUID.randomUUID();
-
-        logger.info("Send message to all with UUID " + offerId);
-        webSocketHandler.sendMessageToAll("Offer " + offerId + " bought!");
     }
 }
