--- offer-provider-service/src/main/java/org/microarchitecturovisco/offerprovider/websockets/OfferProviderWebSocketHandler.java
@@ -1,45 +1,0 @@
-package org.microarchitecturovisco.offerprovider.websockets;
-
-import org.springframework.stereotype.Component;
-import org.springframework.web.socket.CloseStatus;
-import org.springframework.web.socket.TextMessage;
-import org.springframework.web.socket.WebSocketSession;
-import org.springframework.web.socket.handler.TextWebSocketHandler;
-
-import java.io.IOException;
-import java.util.concurrent.CopyOnWriteArrayList;
-
-@Component
-public class OfferProviderWebSocketHandler extends TextWebSocketHandler {
-
-    private final CopyOnWriteArrayList<WebSocketSession> sessions = new CopyOnWriteArrayList<>();
-
-    @Override
-    public void afterConnectionEstablished(WebSocketSession session) {
-        System.out.println("Adding session " + session);
-        sessions.add(session);
-    }
-
-    @Override
-    public void handleTextMessage(WebSocketSession session, TextMessage message) {
-
-    }
-
-    @Override
-    public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
-        sessions.remove(session);
-        super.afterConnectionClosed(session, status);
-    }
-
-    public void sendMessageToAll(String message) {
-        for (WebSocketSession session : sessions) {
-            try {
-                if (session.isOpen()) {
-                    session.sendMessage(new TextMessage(message));
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-}
