+++ reservation-service/src/main/java/org/microarchitecturovisco/reservationservice/services/ReservationService.java
@@ -24,6 +24,7 @@ import org.microarchitecturovisco.reservationservice.services.saga.BookHotelsSag
 import org.microarchitecturovisco.reservationservice.services.saga.BookTransportsSaga;
 import org.microarchitecturovisco.reservationservice.utils.json.JsonConverter;
 import org.microarchitecturovisco.reservationservice.utils.json.JsonReader;
+import org.microarchitecturovisco.reservationservice.websockets.ReservationWebSocketHandler;
 import org.slf4j.LoggerFactory;
 import org.springframework.amqp.AmqpException;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -55,6 +56,8 @@ public class ReservationService {
     private final BookTransportsSaga bookTransportsSaga;
 
     private final RabbitTemplate rabbitTemplate;
+
+    private final ReservationWebSocketHandler reservationWebSocketHandler;
 
     public Reservation createReservation(LocalDateTime hotelTimeFrom, LocalDateTime hotelTimeTo,
                                                       int infantsQuantity, int kidsQuantity, int teensQuantity, int adultsQuantity,
@@ -251,6 +254,8 @@ public class ReservationService {
 
         logger.info("Purchased reservation: " + reservation);
 
+        sendBoughtOfferWebsocketMessages("Ktoś kupił wycieczkę do aktualnie przeglądanego hotelu!", String.valueOf(reservation.getHotelId()));
+
         return ReservationConfirmationResponse.builder()
                 .hotelName(hotelInfo.getName())
                 .price(reservation.getPrice())
@@ -263,6 +268,10 @@ public class ReservationService {
                 .roomTypes(hotelInfo.getRoomTypes())
                 .transport(transportInfo)
                 .build();
+    }
+
+    private void sendBoughtOfferWebsocketMessages(String message, String idHotel) {
+        reservationWebSocketHandler.sendMessageToSubscribedByIdHotel(message, idHotel);
     }
 
     private HotelInfo getHotelInformation(UUID hotelId) {
