+++ reservation-service/src/main/java/org/microarchitecturovisco/reservationservice/websockets/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package org.microarchitecturovisco.offerprovider.websockets;
+package org.microarchitecturovisco.reservationservice.websockets;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
@@ -11,10 +11,10 @@ import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry
 @RequiredArgsConstructor
 public class WebSocketConfig implements WebSocketConfigurer {
 
-    private final OfferProviderWebSocketHandler webSocketHandler;
+    private final ReservationWebSocketHandler webSocketHandler;
 
     @Override
     public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-        registry.addHandler(webSocketHandler, "/offers/ws/offerBought").setAllowedOrigins("*");
+        registry.addHandler(webSocketHandler, "/reservations/ws/offerBought").setAllowedOrigins("*");
     }
 }
